+++ packages/x-date-pickers/src/AdapterDateFnsJalaliV3/index.ts
@@ -1,1 +1,1 @@
-export { AdapterDateFnsJalali } from './AdapterDateFnsJalali';
+export { AdapterDateFnsJalali } from './AdapterDateFnsJalaliV3';
